apps/document-ingestion/requirements.txt
# Production dependencies
# Alternative to Poetry - use this if not using Poetry for dependency management
# To install: pip install -r requirements.txt

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Message Queue
pika==1.3.2  # RabbitMQ client
aio-pika==9.2.1  # Async RabbitMQ client (recommended)

# Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Document Parsing
pypdf2==3.0.1  # PDF parsing
python-docx==1.1.0  # DOCX parsing

# Text Processing
tiktoken==0.5.2  # Token counting

# Embeddings
openai==1.12.0  # Azure OpenAI client

# Vector Store
qdrant-client==1.8.0

# HTTP Client
httpx==0.25.2

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Retry Logic
tenacity==8.2.3

# Utilities
python-json-logger==2.0.7

